ugc/requirements.txt
pydantic[dotenv]==1.10.2
asyncio==3.4.3
aiokafka==0.8.0
aiohttp==3.8.3
orjson==3.8.2
uvicorn==0.20.0
gunicorn==20.1.0
fastapi-utils==0.2.1
fastapi-contrib==0.2.11
clickhouse-driver==0.2.5

sentry-sdk[fastapi]==1.15.0
python-logstash-async==2.5.0